Refactor Makefile and update AudioSettings, MonitoringController, and GeneralSettingsView for improved configuration and UI responsiveness

+++ Sources/CamillaDSPMonitor/Models/AudioSettings.swift
@@ -73,13 +73,13 @@ final class AudioSettings {
   var isMuted: Bool = false {
     didSet { defaults.set(isMuted, forKey: "isMuted") }
   }
-  var silenceThreshold: Double = -60.0 {
+  var silenceThreshold: Int = -60 {
     didSet {
       defaults.set(silenceThreshold, forKey: "silenceThreshold")
       onChanged?()
     }
   }
-  var silenceTimeout: Double = 0.0 {
+  var silenceTimeout: Int = 0 {
     didSet {
       defaults.set(silenceTimeout, forKey: "silenceTimeout")
       onChanged?()
@@ -99,8 +99,8 @@ final class AudioSettings {
     enableRateAdjust = defaults.bool(forKey: "enableRateAdjust")
     resamplerEnabled = defaults.bool(forKey: "resamplerEnabled")
 
-    silenceThreshold = defaults.object(forKey: "silenceThreshold") as? Double ?? -60.0
-    silenceTimeout = defaults.object(forKey: "silenceTimeout") as? Double ?? 0.0
+    silenceThreshold = defaults.object(forKey: "silenceThreshold") as? Int ?? -60
+    silenceTimeout = defaults.object(forKey: "silenceTimeout") as? Int ?? 0
 
     if let t = defaults.string(forKey: "resamplerType"), let type = ResamplerType(rawValue: t) {
       resamplerType = type
